+++ py_scripts/decode.py
