+++ py_scripts/extract_coding_chs.py
